phase2/requirements.txt
# Phase 1 deps
anthropic>=0.25.0
openai-whisper>=20231117
pyttsx3>=2.90
pyaudio>=0.2.13
keyboard>=0.13.5
numpy>=1.24.0
colorama>=0.4.6

# Phase 2 additions
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyautogui>=0.9.54
pillow>=10.0.0
playwright>=1.40.0
